refactor(main): extract shelf, public-account articles and chapters
The three biggest remaining feature blocks move out of main.lua:

- lib/shelf.lua: shelf listing with sort/filter, book details, store
  search and the current-book details view. It owns the translated labels
  for the sort/filter keys handled by lib/shelf_sort.lua, and the
  shelf-page cache marks (refreshCacheIndicators, called by the
  downloader and by cache cleanup).
- lib/mp_articles.lua: public-account article lists, their cache in the
  book record, and downloading a single article.
- lib/chapters.lua: catalog loading (cache then network), the chapter
  list, opening a chapter cached-or-downloaded, the download entry points
  with their annotation-cost warning, and end-of-book handling including
  the ReaderStatus:onEndOfBook hook, now installed and removed through
  installEndOfBookHook / removeEndOfBookHook.

Cross-controller calls go through the plugin instance (self.plugin.shelf,
self.plugin.chapters, ...) so construction order does not matter: every
such call happens inside a callback, long after init() finishes.

Bodies were moved verbatim; only call paths changed. i18n keys unchanged.

main.lua: 2253 -> 1270 lines.
