+++ lib/shelf.lua
@@ -1,0 +1,603 @@
+-- The WeRead bookshelf: shelf listing with sort/filter, per-book details, store
+-- search, and the details view for the currently open book.
+--
+-- Public-account entries are split out of the shelf here and handed to
+-- lib/mp_articles.lua; chapter listing and downloading belong to
+-- lib/chapters.lua. Ordering and filtering rules live in lib/shelf_sort.lua,
+-- with this module owning only their translated labels.
+--
+-- The shelf list carries a cache mark per book. Rebuilding it must stay cheap,
+-- so a downloaded-state cache is threaded through one page build, and
+-- refreshCacheIndicators() re-runs the same builder after a download or a cache
+-- purge instead of re-fetching the shelf.
+
+local ButtonDialog = require("ui/widget/buttondialog")
+local InfoMessage = require("ui/widget/infomessage")
+local InputDialog = require("ui/widget/inputdialog")
+local UIManager = require("ui/uimanager")
+local logger = require("logger")
+local T = require("ffi/util").template
+
+local BookIndex = require("lib.book_index")
+local Content = require("lib.content")
+local I18n = require("lib.i18n")
+local ShelfSort = require("lib.shelf_sort")
+local Util = require("lib.util")
+local WeRead = require("lib.weread")
+
+local function _(text)
+    return I18n.tr(text)
+end
+
+local LOG_MODULE = "[WeRead]"
+
+local log_error = Util.log_error
+local display_error = Util.display_error
+local file_exists = Util.file_exists
+
+local SHELF_SORT_OPTIONS = {
+    { key = "time_desc", label = _("Last read time (newest first)") },
+    { key = "time_asc",  label = _("Last read time (oldest first)") },
+    { key = "default",   label = _("Default order") },
+    { key = "name_asc",  label = _("Title A-Z") },
+    { key = "name_desc", label = _("Title Z-A") },
+}
+
+local function shelfSortLabel(sort_key)
+    for _i, opt in ipairs(SHELF_SORT_OPTIONS) do
+        if opt.key == sort_key then
+            return opt.label
+        end
+    end
+    return SHELF_SORT_OPTIONS[1].label
+end
+
+local SHELF_FILTER_OPTIONS = {
+    { dim = "reading",  value = "finished",       label = _("Only show finished books"),       short = _("Finished") },
+    { dim = "reading",  value = "unfinished",     label = _("Only show unfinished books"),     short = _("Unfinished") },
+    { dim = "download", value = "downloaded",     label = _("Only show downloaded books"),     short = _("Downloaded") },
+    { dim = "download", value = "not_downloaded", label = _("Only show not-downloaded books"), short = _("Not downloaded") },
+}
+
+local Shelf = {}
+Shelf.__index = Shelf
+
+function Shelf:new(plugin)
+    return setmetatable({
+        plugin = plugin,
+        settings = plugin.settings,
+        client = plugin.client,
+        ui_host = plugin.ui_host,
+    }, self)
+end
+
+function Shelf:shelfFilterSummary()
+    local filters = self.shelf_filters
+    local parts = {}
+    for _i, opt in ipairs(SHELF_FILTER_OPTIONS) do
+        if filters[opt.dim] == opt.value then
+            table.insert(parts, opt.short)
+        end
+    end
+    if #parts == 0 then
+        return _("All")
+    end
+    return table.concat(parts, " / ")
+end
+
+function Shelf:saveShelfFilters()
+    local shelf = self.settings:get("shelf")
+    shelf.filter_reading = self.shelf_filters.reading
+    shelf.filter_download = self.shelf_filters.download
+    self.settings:set("shelf", shelf)
+    self.settings:flush()
+end
+
+function Shelf:bookMatchesFilters(book, saved_books, downloaded_cache)
+    return ShelfSort.matches_filters(book, self.shelf_filters, function(candidate)
+        return self:isBookDownloaded(candidate, saved_books, downloaded_cache)
+    end)
+end
+
+function Shelf:isBookDownloaded(book, saved_books, downloaded_cache)
+    return BookIndex.is_downloaded(
+        book, saved_books or self.settings:get("books", {}), downloaded_cache)
+end
+
+function Shelf:showShelfSortOptions(on_sorted)
+    local dialog
+    local current_sort = self.settings:get("shelf").sort_order or "default"
+    local buttons = {}
+    for _i, opt in ipairs(SHELF_SORT_OPTIONS) do
+        table.insert(buttons, {
+            {
+                text = opt.label,
+                checked_func = function()
+                    return opt.key == current_sort
+                end,
+                -- Defer close+refresh so Button's post-tap checkmark repaint runs
+                -- against the still-shown dialog (avoids a ghost label on close).
+                callback = function()
+                    UIManager:nextTick(function()
+                        UIManager:close(dialog)
+                        local shelf = self.settings:get("shelf")
+                        shelf.sort_order = opt.key
+                        self.settings:set("shelf", shelf)
+                        self.settings:flush()
+                        on_sorted()
+                    end)
+                end,
+            },
+        })
+    end
+    dialog = ButtonDialog:new{
+        title = _("Sort by"),
+        title_align = "center",
+        buttons = buttons,
+    }
+    UIManager:show(dialog)
+end
+
+function Shelf:showShelfFilterOptions(on_changed)
+    local dialog
+    local filters = self.shelf_filters
+    local buttons = {
+        {
+            {
+                text = _("All"),
+                checked_func = function()
+                    return filters.reading == nil and filters.download == nil
+                end,
+                callback = function()
+                    UIManager:nextTick(function()
+                        UIManager:close(dialog)
+                        filters.reading = nil
+                        filters.download = nil
+                        self:saveShelfFilters()
+                        on_changed()
+                    end)
+                end,
+            },
+        },
+    }
+    for _i, opt in ipairs(SHELF_FILTER_OPTIONS) do
+        table.insert(buttons, {
+            {
+                text = opt.label,
+                checked_func = function()
+                    return filters[opt.dim] == opt.value
+                end,
+                callback = function()
+                    UIManager:nextTick(function()
+                        UIManager:close(dialog)
+                        -- Toggle within the dimension: re-tapping clears it, else select.
+                        filters[opt.dim] = (filters[opt.dim] == opt.value) and nil or opt.value
+                        self:saveShelfFilters()
+                        on_changed()
+                    end)
+                end,
+            },
+        })
+    end
+    dialog = ButtonDialog:new{
+        title = _("Filter by"),
+        title_align = "center",
+        buttons = buttons,
+    }
+    UIManager:show(dialog)
+end
+
+function Shelf:shelfToolbarItems(with_filters, refresh)
+    local sort_order = self.settings:get("shelf").sort_order
+    local items = {
+        {
+            text = _("Sort"),
+            mandatory = T(_("%1 \u{25BE}"), shelfSortLabel(sort_order)),
+            callback = self.ui_host:safeCallback(_("Sort"), function()
+                self:showShelfSortOptions(refresh)
+            end),
+        },
+    }
+    if with_filters then
+        table.insert(items, {
+            text = _("Filter"),
+            mandatory = T(_("%1 \u{25BE}"), self:shelfFilterSummary()),
+            callback = self.ui_host:safeCallback(_("Filter"), function()
+                self:showShelfFilterOptions(refresh)
+            end),
+        })
+    end
+    items[#items].separator = true -- divide the toolbar rows from the book list
+    return items
+end
+
+function Shelf:showBookshelf()
+    if not self.plugin.account:requireLogin(true, true) then
+        return
+    end
+    self.ui_host:showBusy(_("Loading bookshelf..."))
+    self.ui_host:runOnlineTask(_("Bookshelf"), function()
+        local ok, result = pcall(function()
+            return self.client:gateway("/shelf/sync", {})
+        end)
+        if not ok then
+            self.ui_host:closeBusy()
+            logger.err(LOG_MODULE, "load bookshelf failed:", log_error(result))
+            self.ui_host:showInfo(T(_("Load bookshelf failed:\n%1"), display_error(result)))
+            return
+        end
+        local all_books = result.books or {}
+        local shelf = self.settings:get("shelf")
+        self.shelf_filters = { reading = shelf.filter_reading, download = shelf.filter_download }
+        self.shelf_regular = {}
+        self.shelf_mp = {}
+        for _i, book in ipairs(all_books) do
+            if WeRead.is_mp_book(book.bookId) then
+                table.insert(self.shelf_mp, book)
+            else
+                table.insert(self.shelf_regular, book)
+            end
+        end
+        self.shelf_books = self.shelf_regular
+        self.ui_host:closeBusy()
+        if #self.shelf_mp > 0 then
+            self:showShelfTabs()
+        else
+            self:showShelfPage()
+        end
+    end)
+end
+
+function Shelf:showShelfPage()
+    local books = self.shelf_books or {}
+    if #books == 0 then
+        self.ui_host:showInfo(_("Your WeRead shelf is empty."))
+        return
+    end
+    local menu, buildItems
+    local function refresh()
+        menu:switchItemTable(nil, buildItems())
+    end
+    buildItems = function()
+        local items = self:shelfToolbarItems(true, refresh)
+        local sorted = ShelfSort.sort_books(books, self.settings:get("shelf").sort_order)
+        local saved_books = self.settings:get("books", {})
+        local downloaded_cache = {}
+        self._shelf_saved_books = saved_books
+        for _i, book in ipairs(sorted) do
+            if self:bookMatchesFilters(book, saved_books, downloaded_cache) then
+                local book_id = book.book_id or book.bookId
+                local is_cached = self:isBookDownloaded(book, saved_books, downloaded_cache)
+                local right_text
+                if book.readUpdateTime and book.readUpdateTime > 0 then
+                    right_text = os.date("%Y-%m-%d", book.readUpdateTime)
+                elseif book.finishReading == 1 then
+                    right_text = _("Done")
+                else
+                    right_text = ""
+                end
+                local function rightStatus(cached)
+                    if cached then
+                        return right_text ~= "" and "✓  " .. right_text or "✓"
+                    end
+                    return right_text
+                end
+                table.insert(items, {
+                    text = book.title or book.bookId or _("Untitled"),
+                    mandatory = rightStatus(is_cached),
+                    mandatory_func = function()
+                        local current = self._shelf_saved_books and self._shelf_saved_books[book_id]
+                        return rightStatus(current and file_exists(current.cached_file))
+                    end,
+                    callback = self.ui_host:safeCallback(book.title or book.bookId or _("Untitled"), function()
+                        self:showBookRecord(book)
+                    end),
+                })
+            end
+        end
+        return items
+    end
+    menu = self.ui_host:showList(_("WeRead Bookshelf"), buildItems(), _("Your WeRead shelf is empty."))
+    self.shelf_menu = menu
+    self._shelf_refresh = refresh
+end
+
+-- Re-read the books table and repaint the shelf's cache marks. Called after a
+-- download finishes or a cache entry is cleared, so the open shelf page reflects
+-- the new state without another shelf request.
+function Shelf:refreshCacheIndicators()
+    self._shelf_saved_books = self.settings:get("books", {})
+    if self.shelf_menu and self._shelf_refresh then
+        local ok, err = pcall(self._shelf_refresh)
+        if not ok then
+            logger.warn(LOG_MODULE, "refresh shelf cache indicators failed:", log_error(err))
+        end
+    end
+end
+
+function Shelf:showBookRecord(book)
+    if not self.plugin.account:requireLogin(true, true) then
+        return
+    end
+    local books = self.settings:get("books", {})
+    local book_id = book.book_id or book.bookId
+    if WeRead.is_mp_book(book_id) then
+        self.plugin.mp_articles:showMPAccount(book)
+        return
+    end
+    if book_id then
+        books[book_id] = books[book_id] or {}
+        books[book_id].book_id = book_id
+        books[book_id].title = book.title
+        books[book_id].author = book.author
+        books[book_id].cover = book.cover
+        books[book_id].updated_at = os.time()
+        self.settings:set("books", books)
+        self.settings:flush()
+    end
+    local saved = books[book_id] or book
+    self.ui_host:showBusy(_("Loading book info..."))
+    self.ui_host:runOnlineTask(_("Book info"), function()
+        local ok, err = pcall(function()
+            local info = self.client:get_book_info(book_id)
+            if info then
+                saved.intro = info.intro
+                saved.publisher = info.publisher
+                saved.isbn = info.isbn
+                saved.wordCount = info.wordCount
+                saved.newRating = info.newRating
+                saved.newRatingCount = info.newRatingCount
+                saved.translator = info.translator
+                saved.categoryName = info.categoryName or info.category
+                books[book_id] = saved
+                self.settings:set("books", books)
+                self.settings:flush()
+            end
+            local progress_result = self.client:get_progress(book_id)
+            if progress_result and progress_result.book then
+                saved.progress = progress_result.book.progress or 0
+            end
+        end)
+        self.ui_host:closeBusy()
+        if not ok then
+            logger.err(LOG_MODULE, "load book info failed:", log_error(err))
+            self.ui_host:showInfo(T(_("%1 failed:\n%2"), _("Book info"), display_error(err)))
+            return
+        end
+        self:showBookMenu(saved)
+    end)
+end
+
+function Shelf:showBookMenu(book)
+    local book_id = book.book_id or book.bookId
+    if type(book.chapters) ~= "table" then
+        Content.load_catalog_cache(self.client, self.settings, book)
+    end
+    local menu, buildItems
+    local function refresh()
+        if menu then
+            menu:switchItemTable(nil, buildItems())
+        end
+    end
+
+    buildItems = function()
+        local items = {}
+
+        if book.author and book.author ~= "" then
+            table.insert(items, { text = _("Author"), mandatory = book.author })
+        end
+        if book.translator and book.translator ~= "" then
+            table.insert(items, { text = _("Translator"), mandatory = book.translator })
+        end
+        if book.publisher and book.publisher ~= "" then
+            table.insert(items, { text = _("Publisher"), mandatory = book.publisher })
+        end
+        if book.categoryName and book.categoryName ~= "" then
+            table.insert(items, { text = _("Category"), mandatory = book.categoryName })
+        end
+        if book.wordCount and book.wordCount > 0 then
+            local wc = book.wordCount >= 10000
+                and string.format("%.1f%s", book.wordCount / 10000, _("w words"))
+                or tostring(book.wordCount)
+            table.insert(items, { text = _("Word count"), mandatory = wc })
+        end
+        if book.newRating and book.newRating > 0 then
+            local score = string.format("%.1f", book.newRating / 100)
+            local count = book.newRatingCount and tostring(book.newRatingCount) or "0"
+            table.insert(items, { text = _("Rating"), mandatory = T(_("%1 (%2 ratings)"), score, count) })
+        end
+        if book.isbn and book.isbn ~= "" then
+            table.insert(items, { text = "ISBN", mandatory = book.isbn })
+        end
+        if book.progress and book.progress > 0 then
+            table.insert(items, { text = _("Reading progress"), mandatory = tostring(book.progress) .. "%" })
+        end
+        if book.intro and book.intro ~= "" then
+            table.insert(items, {
+                text = _("Introduction"),
+                callback = function()
+                    UIManager:show(InfoMessage:new{ text = book.intro })
+                end,
+            })
+        end
+
+        if #items > 0 then
+            items[#items].separator = true
+        end
+
+        local saved_books = self.settings:get("books", {})
+        local saved = saved_books[book_id]
+        local cached_path = saved and saved.cached_file or book.cached_file
+        local is_cached = file_exists(cached_path)
+        book.cached_file = is_cached and cached_path or nil
+
+        table.insert(items, {
+            text = _("Chapter list"),
+            post_text = book.chapters and T(_("%1 chapters"), tostring(#book.chapters)) or _("Not loaded"),
+            callback = self.ui_host:safeCallback(_("Chapter list"), function()
+                self.plugin.chapters:showChapterList(book)
+            end),
+        })
+        if is_cached then
+            table.insert(items, {
+                text = _("Clear book cache"),
+                callback = self.ui_host:safeCallback(_("Clear book cache"), function()
+                    self.plugin.cache_admin:confirmClearBookCache(book_id, book.title or book_id, function()
+                        book.cached_file = nil
+                        book.cached_chapters = nil
+                        book.cache_dir = nil
+                        book.chapters = nil
+                        refresh()
+                    end)
+                end),
+            })
+        end
+        table.insert(items, {
+            text = _("Open cached book"),
+            post_text = is_cached and _("Cached") or _("Not cached"),
+            enabled_func = function() return is_cached end,
+            callback = self.ui_host:safeCallback(_("Open cached book"), function()
+                self.plugin.chapters:openCachedBook(book)
+            end),
+        })
+        table.insert(items, {
+            text = _("Download full book"),
+            post_text = _("EPUB"),
+            callback = self.ui_host:safeCallback(_("Download full book"), function()
+                self.plugin.chapters:confirmDownloadAllChapters(book)
+            end),
+        })
+        return items
+    end
+
+    menu = self.ui_host:showList(book.title or _("Book details"), buildItems(), _("No actions."))
+end
+
+function Shelf:showShelfTabs()
+    local items = {
+        {
+            text = _("Books"),
+            post_text = T(_("%1 books"), tostring(#self.shelf_regular)),
+            callback = self.ui_host:safeCallback(_("Books"), function()
+                self.shelf_books = self.shelf_regular
+                self:showShelfPage()
+            end),
+        },
+        {
+            text = _("Public Accounts"),
+            post_text = T(_("%1 accounts"), tostring(#self.shelf_mp)),
+            callback = self.ui_host:safeCallback(_("Public Accounts"), function()
+                self:showMPShelfPage()
+            end),
+        },
+    }
+    self.ui_host:showList(_("WeRead Bookshelf"), items, _("Your WeRead shelf is empty."))
+end
+
+function Shelf:showMPShelfPage()
+    local books = self.shelf_mp or {}
+    if #books == 0 then
+        self.ui_host:showInfo(_("No items."))
+        return
+    end
+    local menu, buildItems
+    local function refresh() menu:switchItemTable(nil, buildItems()) end
+    buildItems = function()
+        local items = self:shelfToolbarItems(false, refresh)
+        local sorted = ShelfSort.sort_books(books, self.settings:get("shelf").sort_order)
+        for _i, book in ipairs(sorted) do
+            table.insert(items, {
+                text = book.title or book.bookId or _("Untitled"),
+                post_text = book.author or "",
+                callback = self.ui_host:safeCallback(book.title or book.bookId or _("Untitled"), function()
+                    self.plugin.mp_articles:showMPAccount(book)
+                end),
+            })
+        end
+        return items
+    end
+    menu = self.ui_host:showList(_("Public Accounts"), buildItems(), _("No items."))
+end
+
+function Shelf:showSearch()
+    if not self.plugin.account:requireLogin(true, true) then
+        return
+    end
+    local dialog
+    dialog = InputDialog:new{
+        title = _("Search WeRead"),
+        input = "",
+        input_type = "text",
+        buttons = {
+            {
+                {
+                    text = _("Cancel"),
+                    id = "close",
+                    callback = self.ui_host:safeCallback(_("Cancel"), function()
+                        UIManager:close(dialog)
+                    end),
+                },
+                {
+                    text = _("Search"),
+                    is_enter_default = true,
+                    callback = self.ui_host:safeCallback(_("Search"), function()
+                        local keyword = dialog:getInputText()
+                        UIManager:close(dialog)
+                        self:searchWithUI(keyword)
+                    end),
+                },
+            },
+        },
+    }
+    self.ui_host:showInputDialog(dialog)
+end
+
+function Shelf:searchWithUI(keyword)
+    if not keyword or keyword == "" then
+        return
+    end
+    self.ui_host:runOnlineTask(_("Search"), function()
+        local ok, result = pcall(function()
+            return self.client:gateway("/store/search", {
+                keyword = keyword,
+                count = 10,
+            })
+        end)
+        if not ok then
+            logger.err(LOG_MODULE, "search failed:", log_error(result))
+            self.ui_host:showInfo(T(_("Search failed:\n%1"), display_error(result)))
+            return
+        end
+        local items = {}
+        for group_index, group in ipairs(result.results or {}) do
+            for book_index, entry in ipairs(group.books or {}) do
+                local book = entry.bookInfo or entry
+                table.insert(items, {
+                    text = book.title or book.bookId or _("Untitled"),
+                    post_text = book.author or "",
+                    mandatory = book.category or "",
+                    callback = self.ui_host:safeCallback(book.title or book.bookId or _("Untitled"), function()
+                        self:showBookRecord(book)
+                    end),
+                })
+            end
+        end
+        self.ui_host:showList(T(_("Search: %1"), keyword), items, _("No search results."))
+    end)
+end
+
+function Shelf:showCurrentBookDetails()
+    if not self.plugin.account:requireLogin(true, true) then
+        return
+    end
+    local book_id = self.plugin:detectWeReadBook()
+    local book = book_id and self.settings:get("books", {})[book_id] or nil
+    if not book then
+        self.ui_host:showInfo(_("The current document is not a WeRead cached book."))
+        return
+    end
+    book.book_id = book.book_id or book_id
+    self:showBookRecord(book)
+end
+
+return Shelf
